(ui) improved limitation add/edit/destroy

+++ app/frontend/src/javascript/components/base/edit-destroy-buttons.tsx
@@ -12,6 +12,7 @@ interface EditDestroyButtonsProps {
   itemId: number,
   itemType: string,
   apiDestroy: (itemId: number) => Promise<void>,
+  confirmationTitle?: string,
   confirmationMessage?: string|ReactNode,
   className?: string,
   iconSize?: number,
@@ -23,7 +24,7 @@ interface EditDestroyButtonsProps {
  * Destroy : shows a modal dialog to ask the user for confirmation about the deletion of the provided item.
  * Edit : triggers the provided function.
  */
-export const EditDestroyButtons: React.FC<EditDestroyButtonsProps> = ({ onDeleteSuccess, onError, onEdit, itemId, itemType, apiDestroy, confirmationMessage, className, iconSize = 20, showEditButton = true }) => {
+export const EditDestroyButtons: React.FC<EditDestroyButtonsProps> = ({ onDeleteSuccess, onError, onEdit, itemId, itemType, apiDestroy, confirmationTitle, confirmationMessage, className, iconSize = 20, showEditButton = true }) => {
   const { t } = useTranslation('admin');
 
   const [deletionModal, setDeletionModal] = useState<boolean>(false);
@@ -58,7 +59,7 @@ export const EditDestroyButtons: React.FC<EditDestroyButtonsProps> = ({ onDelete
           <Trash size={iconSize} weight="fill" />
         </FabButton>
       </div>
-      <FabModal title={t('app.admin.edit_destroy_buttons.delete_item', { TYPE: itemType })}
+      <FabModal title={confirmationTitle || t('app.admin.edit_destroy_buttons.delete_item', { TYPE: itemType })}
         isOpen={deletionModal}
         toggleModal={toggleDeletionModal}
         closeButton={true}
